requirements.txt
flask
flask-cors
dotenv
google-generativeai